tst-timespec.c: Explicitly cast TIME_T_MAX to double
Explicitly cast TIME_T_MAX to double in tst-timespec.c to silence Clang
error:

tst-timespec.c:290:19: error: implicit conversion from 'time_t' (aka 'long') to 'double' changes value from 9223372036854775807 to 9223372036854775808 [-Werror,-Wimplicit-const-int-float-conversion]
  287 |   {.expected = {.tv_sec = 0, .tv_nsec = 1},
      |   ~
  288 |    .observed = {.tv_sec = TIME_T_MAX / TIMESPEC_HZ,
  289 |                 .tv_nsec = TIMESPEC_HZ - 1},
  290 |    .upper_bound = TIME_T_MAX, .lower_bound = 1, .result = 1,
      |                   ^~~~~~~~~~

+++ support/tst-timespec.c
@@ -287,7 +287,7 @@ struct timespec_test_case check_cases[] = {
   {.expected = {.tv_sec = 0, .tv_nsec = 1},
    .observed = {.tv_sec = TIME_T_MAX / TIMESPEC_HZ,
 		.tv_nsec = TIMESPEC_HZ - 1},
-   .upper_bound = TIME_T_MAX, .lower_bound = 1, .result = 1,
+   .upper_bound = (double) TIME_T_MAX, .lower_bound = 1, .result = 1,
   },
 };
